Implement HEAD
Not sure if anything will use HEAD, but I realized that FastAPI
doesn't implement it by default and there is substantial optimization
possible in HEAD vs. GET since the content doesn't have to be
retrieved from GCS.  Implement it, just in case, and refactor the
handler to separate out common code.