Redirect URLs with duplicate slashes
It's easy to accidentally get duplicate slashes into URLs by, for
instance, configuring a URL prefix with a trailing slash and then
adding paths with a leading slash to it.  Normal web servers
transparently deal with such URLs, but crawlspace was rejecting
them with 422 errors.

Broaden the regex to accept them and then detect them and force a
permanent redirect to the cleaned-up URL, which copies the standard
behavior of a regular web server.