Fix quoting of Etag header
The etag value returned by the GCS API doesn't include the double
quotes that are required in the Etag header format, which in turn
led browsers to send the same value back, at which point we'd
reject it because the syntax wasn't valid.  Add the double quotes
and test it properly.